# Add function barrier in `rand!`