This PR is an alternative to https://github.com/JuliaStats/Distributions.jl/pull/1281 that does not introduce any additional keyword arguments for `rand!`.

It introduces a function barrier in `rand!(rng, s::Sampleable{Univariate}, ::AbstractArray)` since `sampler(s)` is not type stable e.g. for `Beta` distributions.

On master:
```julia
julia> using BenchmarkTools, Distributions, Random

julia> x = zeros(100_000);

julia> f(x) = rand!(Beta(0.1, 1.0), x);

julia> @benchmark f($x)
BenchmarkTools.Trial: 
  memory estimate:  3.04 MiB
  allocs estimate:  199490
  --------------
  minimum time:     18.404 ms (0.00% GC)
  median time:      19.938 ms (0.00% GC)
  mean time:        20.054 ms (0.38% GC)
  maximum time:     23.241 ms (0.00% GC)
  --------------
  samples:          249
  evals/sample:     1
```

With this PR:
```julia
julia> using BenchmarkTools, Distributions, Random

julia> x = zeros(100_000);

julia> f(x) = rand!(Beta(0.1, 1.0), x);

julia> @benchmark f($x)
BenchmarkTools.Trial: 
  memory estimate:  64 bytes
  allocs estimate:  1
  --------------
  minimum time:     13.786 ms (0.00% GC)
  median time:      14.192 ms (0.00% GC)
  mean time:        14.275 ms (0.00% GC)
  maximum time:     17.123 ms (0.00% GC)
  --------------
  samples:          350
  evals/sample:     1
```